adjust test case
