+++ tests/iceberg/test_model.py
@@ -1,4 +1,3 @@
-from pydantic import BaseModel
 import pytest
 from pyiceberg.types import (
     IntegerType,
@@ -83,18 +82,6 @@ class TestIcebergFastDataframeModel:
         assert isinstance(field.field_type, expected_iceberg_type)
         assert field.required is expected_required
 
-    def test_model_conversion(self) -> None:
-        class BaseTransaction(BaseModel):
-            transaction_id: str
-
-        class Transaction(BaseTransaction):
-            amount: float
-            timestamp: str
-
-        # Convert to Iceberg schema
-        IcebergTransaction = IcebergFastDataframeModel.from_base_model(Transaction)
-        iceberg_schema = IcebergTransaction.iceberg_schema()
-        assert len(iceberg_schema.fields) == 3
 
     class TestIcebergValidation:
         @pytest.fixture
